Move coincidence_type_str_to_in to a class method of CoincidenceBitmap Move CoincidenceBitmap to the hi.uitils module Move tests to appropriate files
