+++ imap_processing/hi/l1c/hi_l1c.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import logging
-import re
 from pathlib import Path
 
 import numpy as np
@@ -13,8 +12,11 @@ import xarray as xr
 from imap_processing.cdf.imap_cdf_manager import ImapCdfAttributes
 from imap_processing.cdf.utils import parse_filename_like
 from imap_processing.hi.l1a.science_direct_event import DE_CLOCK_TICK_S
-from imap_processing.hi.l1b.hi_l1b import CoincidenceBitmap
-from imap_processing.hi.utils import create_dataset_variables, full_dataarray
+from imap_processing.hi.utils import (
+    CoincidenceBitmap,
+    create_dataset_variables,
+    full_dataarray,
+)
 from imap_processing.spice.geometry import (
     SpiceFrame,
     frame_transform,
@@ -478,7 +480,7 @@ class CalibrationProductConfig:
         # to integer values
         self._obj["coincidence_type_values"] = self._obj.apply(
             lambda row: [
-                coincidence_type_string_to_int(entry)
+                CoincidenceBitmap.detector_hit_str_to_int(entry)
                 for entry in row["coincidence_type_list"]
             ],
             axis=1,
@@ -521,33 +523,3 @@ class CalibrationProductConfig:
             calibration product definitions.
         """
         return len(self._obj.index.unique(level="cal_prod_num"))
-
-
-def coincidence_type_string_to_int(coincidence_type_str: str) -> int:
-    """
-    Convert a coincidence type string to a coincidence type integer value.
-
-    A coincidence string is a string containing all detectors that were hit
-    for a direct event. Possible detectors include: [A, B, C1, C2]. Converting
-    the coincidence type string to a coincidence type integer value involves
-    summing the coincidence bitmap value for each detector hit. e.g. "AC1C2"
-    results in 2**3 + 2**1 + 2**0 = 11. See `CoincidenceBitmap` for the mapping
-    from detector name to integer value.
-
-    Parameters
-    ----------
-    coincidence_type_str : str
-        The coincidence type string containing the list of detectors hit.
-        e.g. "AC1C2".
-
-    Returns
-    -------
-    coincidence_type : int
-        The integer value of the coincidence type.
-    """
-    # CoincidenceBitmap defines the detector names and their associated
-    # values. Use regex to find matches to the detector names.
-    pattern = r"|".join(c.name for c in CoincidenceBitmap)
-    matches = re.findall(pattern, coincidence_type_str)
-    # Sum the integer value assigned to the detector name for each match
-    return sum(CoincidenceBitmap[m] for m in matches)
